Move `inputWidget` types
